Cleanup for clarity on interactors.
Several of the changes here are more OneIdeaPerLine and cleanups for
high flog scores. But *still* this makes almost no dent in the total
flog of the application. We're mostly shuffling complexity around.

However, also added default() and validate() calls to the ServiceObject
class, because a huge amount of each class has been dealing with the
setup of those objects.

I'd still like to do something to eliminate the scourge of `context.foo`
calls all over the interactions, but that will have to wait.